Make the `_toggle_quotes` method more efficient.
This is all new to me so... per ChatGPT:

You can use the translate method of strings in python to make the
function more readable and efficient.

```Python
def _toggle_quotes(text):
    return text.translate(str.maketrans("'\"", '"\''))
```

This method creates a translation table using the str.maketrans function
and then applies it to the text using the translate method, effectively
replacing all occurrences of single quotes with double quotes and vice
versa. This method is more readable and also more efficient than the
previous examples.