Reorganize code
I believe this is the more idiomatic Python way to define a utility
method that is only needed withing the scope of a class and that doesn't
need access to state. Wasn't sure how to do this at first so I had
moved the method up and out of the scope of the class definition to
make it easily callable (without using `self.`, for example), but that
just doesn't seem to be how Python works.